added map and job seeker page
I also made some visual changes to the layout of the login page, and got
the map on the page side-by-side to the login area.
For the job seeker page, I got the functionality of the page, which is
all job seekers that log in will see all of the job postings in the
database.