aftermath of merge
fixed some of the issues that occured as a result of today's merge of
all of our projects.
loginview.php had alot of html probs that we fixed.
added javascript folders with scripts inside for better organization
changed database.php password to be empty string so that i could log in.